server/requirements.txt
fastapi==0.123.0
uvicorn==0.38.0
PyMuPDF>=1.26.0
pymupdf4llm==0.2.2
pdfplumber==0.11.0
python-multipart>=0.0.12
pydantic==2.12.5
pydantic-settings>=2.4.0

# LangChain & LangGraph
langchain==1.1.0
langgraph==1.0.4
langchain-core==1.1.0
langchain-community==0.4.1
langchain-google-genai==3.2.0
langchain-text-splitters
langchain-chroma

# Google Cloud
google-cloud-storage==3.6.0
google-genai==1.52.0

# Database
python-dotenv
chromadb